Refactor file namesto use CamelCaseUppercase, remove useless Onsen component files
